+++ CHANGELOG.md
@@ -6,7 +6,7 @@ Categories for each release: Added, Changed, Deprecated, Removed, Fixed, Securit
 
 ## [Unreleased]
 
-* `dx get` allows to create a local representation of a workflow
+* `dx get workflow-xxxx` allows to create a local representation of a workflow
 
 ## [215.0] - beta